feature: remove existing enclaves of same name and save on endpoint deploy
