+++ src/commands/rollup-deploy.ts
@@ -1,11 +1,11 @@
 import fs from 'fs';
 import yaml from 'js-yaml';
-import { getProjectName } from '../configs/project';
+import { getProjectName ,getOverwriteExistingEnclave as getOverwriteExistingEnclave} from '../configs/project';
 import { cloneOptimismPacakge } from '../utils/clone';
 import { WEBSITE } from '../utils/config';
 import { runKurtosisCommand } from '../utils/system';
-import { deployFailedLog, rollupConfigLog } from '../utils/log';
-import { ensureProjectDirectory, getProjectConfig, removeExistingEnclave } from '../utils/project-manage';
+import { deployFailedLog, rollupConfigLog, deployCompleteLog} from '../utils/log';
+import { ensureProjectDirectory, getProjectConfig, overwriteExistingEnclave, saveChainInfo,  } from '../utils/project-manage';
 
 export async function RollupDeploy(options: {id?: string, file?: string}) {
   const {id, file} = options;
@@ -18,7 +18,12 @@ export async function RollupDeploy(options: {id?: string, file?: string}) {
 
   // make a directory with the project name in a project folder
   await ensureProjectDirectory(projectName.projectName);
-  await removeExistingEnclave(projectName.projectName);
+
+  const {removeEnclave} = await getOverwriteExistingEnclave();
+  if (removeEnclave){
+    await overwriteExistingEnclave(projectName.projectName);
+  }
+
 
   try{
     if (id){
@@ -32,6 +37,9 @@ export async function RollupDeploy(options: {id?: string, file?: string}) {
     else{
       throw new Error("Either --id or --file must be specified");
     }
+
+    await saveChainInfo(projectName.projectName);
+    deployCompleteLog();
   }
   catch(error){
     deployFailedLog(String(error));
@@ -39,10 +47,8 @@ export async function RollupDeploy(options: {id?: string, file?: string}) {
 
 }
 
-
 export async function RollupDeployFile(file_path:string, projectName: string) {
   // try to read the file and save it to a yaml in the project folder
-  try{
     if (!file_path.endsWith('.yaml')){
       throw new Error('File must be a yaml file');
     }
@@ -56,13 +62,10 @@ export async function RollupDeployFile(file_path:string, projectName: string) {
     await fs.promises.writeFile(configPath, yamlString);
     console.log('Saved config file');
     await deployFromConfig(configPath, projectName);
-  }catch(error){  
-    deployFailedLog(String(error));
   }
-}
 
 export async function RollupDeployEndpoint(id:string, projectName: string) {
-  try{
+
     const endpoint = WEBSITE.ENDPOINT.replace('[id]', id);
     
     
@@ -71,9 +74,6 @@ export async function RollupDeployEndpoint(id:string, projectName: string) {
     console.log('Downloaded config file');
     await deployFromConfig(configPath, projectName);
  
-  } catch(error){
-    deployFailedLog(String(error));
-  }
 }
 
 async function downloadConfig(endpoint:string, configPath:string) {  
@@ -83,10 +83,8 @@ async function downloadConfig(endpoint:string, configPath:string) {
       throw new Error('Could not download config file');
     }
     const text = await response.text();
-    const yamltext = yaml.load(text);
-    const yamlObj = yaml.dump(yamltext);
 
-    await fs.promises.writeFile(configPath, yamlObj);
+    await fs.promises.writeFile(configPath, text);
 }
 
 async function deployFromConfig(configPath:string, projectName:string) {
